Add files via upload
In this task, a serverless function was created and deployed in Google Cloud .The function executes code automatically in response to an event, without managing any servers. Public HTTP triggers were used to test the function, and the deployment demonstrates event-driven computing, resource optimization, and cost-effective cloud operations. This task helps understand serverless architecture, Function-as-a-Service (FaaS), and automated cloud workflows.